refactor: reorganize agent/tools and tests/tools into domain subdirectories
agent/tools/
  garden/     ← profile, plants, beds_containers, search
  projects/   ← projects, planning, tracker
  operations/ ← triage, care, incidents, interactions, weather, activity

tests/tools/
  garden/     ← test_profile, test_plants, test_beds_containers, test_search
  projects/   ← test_projects, test_planning, test_task_tracker_tools,
                 test_priority_and_progress, test_bulk_assign, test_query_efficiency
  operations/ ← test_triage_care_incident_operations, test_activity,
                 test_interaction_tools, test_status_and_orphans

All import paths updated across main.py, agent/tools/__init__.py,
tests/support/patching.py, and all test files. No backwards compatibility shims.
286 tests pass.
