+++ agent/tools/garden/plants.py
